Fix manifest impl
